backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
openai==0.27.8
elevenlabs==0.2.24
python-dotenv==1.0.0
gunicorn==20.1.0
PyPDF2==3.0.1